Update to remove POST
I have removed the POST examples from this document per the request of Ankit. Instead I have provided a sample curl GET request. This is a temporary change. I will add a sample response too, and later hopefully provide an entire guide on all the methods. But this is a quick change so that customers are not confused about POST for this being currently supported.