Style revisions for multiple breakdown reports
For this guide, I provided some organizational headings, formatting and language consistency and some contextual information.

As for the purpose of using APIs for multiple breakdown reports, I had a hard time understanding "cross-product values from differing dimensions" in a concrete way, although it seems an apt way to describe it abstractly. I searched for a use case for using the reports API for multiple breakdowns and the result is the second paragraph I added to this guide.  You may not like it because it is not the example that is provided later in the guide itself or in the Reports User Guide (which is also a separate example).  But I thought it might be helpful to provide a holistic visualization function that answers "why" and not just "how." If you do not like it, we can remove it.