BDIA changes
Doc covers a lot of bases. My only content suggestion to add is a  sample request and response put together that has a good explanation of what is being specified in the request (so an explanation of what is in any batch file too) and then a description of what exactly the JSON response is. We've got the good table defining the response fields, but it might also help to show those in the actual JSON example and then text to explain the values in the field. This might provide some additional context. We've got all of the helpful JSON error responses in there already, so that's great.






For example,