Update flow of example info
Taking @gigazelle advice in part on this and changing my mind. The fact that both request and response example appear at same place in layout make the ordering considerations a bit different than I was used to. As a result, I'm putting the both examples first and then adding parameters. I'm also changing the org a bit in other ways as the parameter section is now a sub-unit of each example details section.